requirements.txt
autopep8==1.5
beautifulsoup4==4.8.2
cassandra-driver==3.23.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
elasticsearch==7.7.1
extruct==0.8.0
geomet==0.1.2
hashids==1.2.0
html-text==0.5.1
html5lib==1.0.1
idna==2.9
isodate==0.6.0
kafka-python==2.0.2
langdetect==1.0.9
lxml==4.5.0
mf2py==1.1.2
nltk==3.4.5
oauthlib==3.1.0
pkg-resources==0.0.0
pybase62==0.4.3
pycld2==0.41
pycodestyle==2.5.0
pyparsing==2.4.6
PyPDF2==1.26.0
PySocks==1.7.1
python-dateutil==2.8.1
pyutil==3.3.0
rdflib==4.2.2
rdflib-jsonld==0.4.0
requests==2.23.0
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.0
textblob==0.15.3
toml==0.10.2
tweepy==3.8.0
urllib3==1.22
w3lib==1.21.0
webencodings==0.5.1
